refactor: move persist_to_session into session module
SQL statement lives with other session DB logic.
